+++ android/src/main/java/cn/qiuxiang/react/amap3d/utils/LatLongData.kt
@@ -2,5 +2,5 @@ package cn.qiuxiang.react.amap3d.utils
 
 class LatLongData(var latitude: Double//维度
                   , var longitude: Double//经度
-                  , var speed: Double//速度
+                  , var speed: Float//速度
 )
